Add Chat Room and Group host search pages

+++ GoTvCore/GoTvP2P/Plugins/PluginBaseRelay.cpp
@@ -149,7 +149,7 @@ EPluginAccessState PluginBaseRelay::handlePktRelayServiceReq( BigListInfo * bigL
 				bool grantedUserSession = false;
 				if( ePluginAccessOk == bigListInfo->getHisAccessPermissionFromMe( m_ePluginType ) )
 				{
-					if( m_FriendGuidList.size() < m_MaxUserRelayAllowCnt )
+					if( (int)m_FriendGuidList.size() < m_MaxUserRelayAllowCnt )
 					{
 						m_FriendGuidList.addGuid( bigListInfo->getMyOnlineId() );
 						grantedUserSession = true;
@@ -160,7 +160,7 @@ EPluginAccessState PluginBaseRelay::handlePktRelayServiceReq( BigListInfo * bigL
 				if( false == grantedUserSession )
 				{
 					// try anonymous list
-					if( m_AnonGuidList.size() < m_MaxAnonRelayAllowCnt )
+					if( ( int )m_AnonGuidList.size() < m_MaxAnonRelayAllowCnt )
 					{
 						m_AnonGuidList.addGuid( bigListInfo->getMyOnlineId() );
 						grantedAnonSession = true;
